Refactor how the testCase output is evaluated
We now match each SubjectAccessReview object with the expected result.
We then populate a list of all those that didn't meet expectation.
If a test fails, we use the String() function on the testCaseOutput
to print all these violating SARs.
If there are not such SARs, we consider the test passed.